importer/requirements.txt
appdirs==1.4.4
application-utility==1.3.2
autopep8==1.7.0
Brlapi==0.8.4
btrfsutil==5.18.1
CacheControl==0.12.6
cachetools==5.2.0
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
colorama==0.4.5
commonmark==0.9.1
cryptography==37.0.4
cupshelpers==1.0
distlib==0.3.5
distro==1.7.0
docopt==0.6.2
docutils==0.19
future==0.18.2
google-api-core==2.8.2
google-auth==2.11.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==1.1.3
idna==3.3
jaraco.context==4.1.1
jaraco.functools==3.5.1
jaraco.text==3.8.1
keyutils==0.6
libfdt==1.6.1
lit==14.0.6.dev0
louis==3.22.0
manjaro-sdk==0.1
more-itertools==8.13.0
msgpack==1.0.3
netsnmp-python==1.0a1
nftables==0.1
npyscreen==4.10.5
nspektr==0.4.0
numpy==1.23.2
ordered-set==4.0.2
packaging==21.3
pacman-mirrors==4.23.2
pandas==1.4.3
pep517==0.12.0
Pillow==9.2.0
platformdirs==2.5.2
ply==3.11
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.21.0
pycodestyle==2.9.1
pycparser==2.21
pycups==2.0.1
pycurl==7.45.1
Pygments==2.12.0
PyGObject==3.42.2
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySide6==6.3.1
pysmbc==1.0.23
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
rados==2.0.0
rbd==2.0.0
reportlab==3.6.11
requests==2.28.1
resolvelib==0.8.1
rgw==2.0.0
rich @ file:///build/python-rich/src/rich-12.5.1/dist/rich-12.5.0-py3-none-any.whl
rsa==4.9
shiboken6==6.3.1
six==1.16.0
SQLAlchemy==1.4.40
systemd-python==234
TBB==0.2
team==1.0
tenacity==8.0.1
toml==0.10.2
tomli==2.0.1
trove-classifiers==2022.8.7
typing_extensions==4.3.0
udiskie==2.4.2
urllib3==1.26.11
validate-pyproject==0.9
webencodings==0.5.1
